refactor: remove unused DTOs and update import paths for ParameterDto
