--- src/core/dtos/CreateParameterDto.ts
@@ -1,8 +1,0 @@
-export type CreateParameterDto = {
-  name: string;
-  unit: string;
-  factor: number;
-  offset: number;
-  active: boolean;
-  iconType?: string;
-};
